// frontend/src/types/profile.ts
export interface CandidateProfile {
  id: number;

  firstName: string;

  lastName: string;

  email: string;

  phone: string;

  profilePicture: string;

  bio: string;

  skills: string;

  education: string;

  experience: string;

  linkedin: string;

  github: string;

  portfolio: string;

  location: string;
}